Put SongAdapter in a separate file
